v001

+++ src/Subscription.php
@@ -6,14 +6,16 @@ use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Event;
-use Laravel\Cashier\Order\Contracts\InteractsWithOrderItems;
-use Laravel\Cashier\Order\Contracts\PreprocessesOrderItems;
 use Laravel\Cashier\Coupon\AppliedCoupon;
 use Laravel\Cashier\Coupon\Contracts\AcceptsCoupons;
 use Laravel\Cashier\Coupon\RedeemedCoupon;
 use Laravel\Cashier\Events\SubscriptionCancelled;
 use Laravel\Cashier\Events\SubscriptionPlanSwapped;
 use Laravel\Cashier\Events\SubscriptionQuantityUpdated;
+use Laravel\Cashier\Events\SubscriptionResumed;
+use Laravel\Cashier\Events\SubscriptionStarted;
+use Laravel\Cashier\Order\Contracts\InteractsWithOrderItems;
+use Laravel\Cashier\Order\Contracts\PreprocessesOrderItems;
 use Laravel\Cashier\Order\Order;
 use Laravel\Cashier\Order\OrderItem;
 use Laravel\Cashier\Order\OrderItemCollection;
@@ -43,7 +45,7 @@ use Money\Money;
  */
 class Subscription extends Model implements InteractsWithOrderItems, PreprocessesOrderItems, AcceptsCoupons
 {
-    use hasOwner;
+    use HasOwner;
 
     /**
      * The attributes that are not mass assignable.
@@ -62,6 +64,15 @@ class Subscription extends Model implements InteractsWithOrderItems, Preprocesse
         'cycle_started_at',
         'cycle_ends_at',
         'ends_at',
+    ];
+
+    /**
+     * The event map for the model.
+     *
+     * @var array
+     */
+    protected $dispatchesEvents = [
+        'created' => SubscriptionStarted::class,
     ];
 
     /**
@@ -145,19 +156,21 @@ class Subscription extends Model implements InteractsWithOrderItems, Preprocesse
     public function getCycleProgressAttribute($now = null, $precision = 5)
     {
         $now = $now ?: now();
+        $cycle_started_at = $this->cycle_started_at->copy();
+        $cycle_ends_at = $this->cancelled() ? $this->ends_at->copy() : $this->cycle_ends_at->copy();
 
-        // completed
-        if($this->cycle_ends_at->lessThanOrEqualTo($now)) {
+        // Cycle completed
+        if ($cycle_ends_at->lessThanOrEqualTo($now)) {
             return 1;
         }
 
-        // not yet started
-        if($this->cycle_started_at->greaterThanOrEqualTo($now)) {
+        // Cycle not yet started
+        if ($cycle_started_at->greaterThanOrEqualTo($now)) {
             return 0;
         }
 
-        $total_cycle_seconds = $this->cycle_started_at->diffInSeconds($this->cycle_ends_at);
-        $seconds_progressed = $this->cycle_started_at->diffInSeconds($now);
+        $total_cycle_seconds = $cycle_started_at->diffInSeconds($cycle_ends_at);
+        $seconds_progressed = $cycle_started_at->diffInSeconds($now);
 
         return round($seconds_progressed / $total_cycle_seconds, $precision);
     }
@@ -185,16 +198,35 @@ class Subscription extends Model implements InteractsWithOrderItems, Preprocesse
      */
     public function swap(string $plan, $invoiceNow = true)
     {
+        /** @var Plan $newPlan */
         $newPlan = app(PlanRepository::class)::findOrFail($plan);
-        $applyNewSettings = function() use ($newPlan) {
+        $previousPlan = $this->plan;
+
+        if ($this->cancelled()) {
+            $this->cycle_ends_at = $this->ends_at;
+            $this->ends_at = null;
+        }
+
+        $applyNewSettings = function () use ($newPlan) {
             $this->plan = $newPlan->name();
         };
 
         $this->restartCycleWithModifications($applyNewSettings, now(), $invoiceNow);
 
-        Event::dispatch(new SubscriptionPlanSwapped($this));
+        Event::dispatch(new SubscriptionPlanSwapped($this, $previousPlan));
 
         return $this;
+    }
+
+    /**
+     * Swap the subscription to a new plan, and invoice immediately.
+     *
+     * @param string $plan
+     * @return $this
+     */
+    public function swapAndInvoice($plan)
+    {
+        return $this->swap($plan, true);
     }
 
     /**
@@ -242,7 +274,7 @@ class Subscription extends Model implements InteractsWithOrderItems, Preprocesse
      *
      * @param \Carbon\Carbon $endsAt
      * @param string $reason
-     * @return mixed
+     * @return $this
      */
     public function cancelAt(Carbon $endsAt, $reason = SubscriptionCancellationReason::UNKNOWN)
     {
@@ -261,6 +293,17 @@ class Subscription extends Model implements InteractsWithOrderItems, Preprocesse
     }
 
     /**
+     * Cancel the subscription immediately.
+     *
+     * @param string $reason
+     * @return $this
+     */
+    public function cancelNow($reason = SubscriptionCancellationReason::UNKNOWN)
+    {
+        return $this->cancelAt(now(), $reason);
+    }
+
+    /**
      * Remove the subscription's scheduled order item.
      * Optionally persists the reference removal on the subscription.
      *
@@ -272,13 +315,13 @@ class Subscription extends Model implements InteractsWithOrderItems, Preprocesse
     {
         $item = $this->scheduledOrderItem;
 
-        if($item->isProcessed(false)) {
+        if ($item && $item->isProcessed(false)) {
             $item->delete();
         }
 
         $this->fill(['scheduled_order_item_id' => null]);
 
-        if($save) {
+        if ($save) {
             $this->save();
         }
 
@@ -310,6 +353,8 @@ class Subscription extends Model implements InteractsWithOrderItems, Preprocesse
                 'ends_at' => null,
                 'scheduled_order_item_id' => $item->id,
             ])->save();
+
+            Event::dispatch(new SubscriptionResumed($this));
 
             return $this;
         });
@@ -357,16 +402,16 @@ class Subscription extends Model implements InteractsWithOrderItems, Preprocesse
      */
     public function scheduleNewOrderItemAt(Carbon $process_at, $item_overrides = [], $fill_link = true, Plan $plan = null)
     {
-        if($this->scheduled_order_item_id)
-        {
+        if ($this->scheduled_order_item_id) {
             throw new LogicException('Cannot schedule a new subscription order item if there is already one scheduled.');
         }
 
-        if(is_null($plan)) {
+        if (is_null($plan)) {
             $plan = $this->plan();
         }
 
-        $item = $this->orderItems()->create(array_merge([
+        $item = $this->orderItems()->create(array_merge(
+            [
                 'owner_id' => $this->owner_id,
                 'owner_type' => $this->owner_type,
                 'process_at' => $process_at,
@@ -374,8 +419,9 @@ class Subscription extends Model implements InteractsWithOrderItems, Preprocesse
                 'unit_price' => (int) $plan->amount()->getAmount(),
                 'quantity' => $this->quantity ?: 1,
                 'tax_percentage' => $this->tax_percentage,
-                'description' => $this->plan()->description(),
-            ], $item_overrides
+                'description' => $plan->description(),
+            ],
+            $item_overrides
         ));
 
         if ($fill_link) {
@@ -412,9 +458,11 @@ class Subscription extends Model implements InteractsWithOrderItems, Preprocesse
         /** @var Subscription scheduled_order_item_id */
         $subscription = $item->orderable;
         $plan_swapped = false;
+        $previousPlan = null;
 
-        if(! empty($subscription->next_plan)) {
+        if (! empty($subscription->next_plan)) {
             $plan_swapped = true;
+            $previousPlan = $subscription->plan;
             $subscription->plan = $subscription->next_plan;
             $subscription->next_plan = null;
         }
@@ -437,8 +485,8 @@ class Subscription extends Model implements InteractsWithOrderItems, Preprocesse
             return $item;
         });
 
-        if($plan_swapped) {
-            Event::dispatch(new SubscriptionPlanSwapped($subscription));
+        if ($plan_swapped) {
+            Event::dispatch(new SubscriptionPlanSwapped($subscription, $previousPlan));
         }
 
         return $item;
@@ -531,6 +579,18 @@ class Subscription extends Model implements InteractsWithOrderItems, Preprocesse
     }
 
     /**
+     * Increment the quantity of the subscription, and invoice immediately.
+     *
+     * @param int $count
+     * @return \Laravel\Cashier\Subscription
+     * @throws \Throwable
+     */
+    public function incrementAndInvoice($count = 1)
+    {
+        return $this->incrementQuantity($count, true);
+    }
+
+    /**
      * Decrement the quantity of the subscription.
      *
      * @param int $count
@@ -560,7 +620,7 @@ class Subscription extends Model implements InteractsWithOrderItems, Preprocesse
 
         $oldQuantity = $this->quantity;
 
-        $this->restartCycleWithModifications(function() use ($quantity) {
+        $this->restartCycleWithModifications(function () use ($quantity) {
             $this->quantity = $quantity;
         }, now(), $invoiceNow);
 
@@ -609,7 +669,7 @@ class Subscription extends Model implements InteractsWithOrderItems, Preprocesse
     protected function reimburseUnusedTime(?Carbon $now = null)
     {
         $now = $now ?: now();
-        if($this->onTrial()) {
+        if ($this->onTrial()) {
             return null;
         }
 
@@ -620,12 +680,12 @@ class Subscription extends Model implements InteractsWithOrderItems, Preprocesse
     }
 
     /**
-     * Wraps up the current billing cycle, applies modifications to this subscription and starts a new cycle.
+     * Wrap up the current billing cycle, apply modifications to this subscription and start a new cycle.
      *
      * @param \Closure $applyNewSettings
      * @param \Carbon\Carbon|null $now
      * @param bool $invoiceNow
-     * @return mixed
+     * @return \Laravel\Cashier\Subscription
      */
     public function restartCycleWithModifications(\Closure $applyNewSettings, ?Carbon $now = null, $invoiceNow = true)
     {
@@ -642,11 +702,10 @@ class Subscription extends Model implements InteractsWithOrderItems, Preprocesse
             // Apply new subscription settings
             call_user_func($applyNewSettings);
 
-            if($this->onTrial()) {
+            if ($this->onTrial()) {
 
                 // Reschedule next cycle's OrderItem using the new subscription settings
                 $orderItems[] = $this->scheduleNewOrderItemAt($this->trial_ends_at);
-
             } else { // Start a new billing cycle using the new subscription settings
 
                 // Reset the billing cycle
@@ -659,13 +718,26 @@ class Subscription extends Model implements InteractsWithOrderItems, Preprocesse
 
             $this->save();
 
-            if($invoiceNow) {
+            if ($invoiceNow) {
                 $order = Order::createFromItems($orderItems);
                 $order->processPayment();
             }
 
             return $this;
         });
+    }
+
+    /**
+     * Wrap up the current billing cycle and start a new cycle.
+     *
+     * @param \Carbon\Carbon|null $now
+     * @param bool $invoiceNow
+     * @return \Laravel\Cashier\Subscription
+     */
+    public function restartCycle(?Carbon $now = null, $invoiceNow = true)
+    {
+        return $this->restartCycleWithModifications(function () {
+        }, $now, $invoiceNow);
     }
 
     /**
